+++ molo/yourwords/models.py
@@ -1,6 +1,6 @@
 from django.db import models
 from django.utils.translation import ugettext_lazy as _
-
+from django.dispatch import receiver
 
 from wagtail.wagtailcore.models import Page
 from wagtail.wagtailcore.fields import StreamField
@@ -12,17 +12,44 @@ from wagtail.wagtailadmin.edit_handlers import (
 from wagtail.wagtailimages.edit_handlers import ImageChooserPanel
 
 from molo.core.blocks import MarkDownBlock
-from molo.core.models import ArticlePage, TranslatablePageMixin, SectionPage
+from molo.core.utils import generate_slug
+from molo.core.models import (
+    ArticlePage,
+    SectionPage,
+    TranslatablePageMixinNotRoutable,
+    PreventDeleteMixin,
+    Main,
+    index_pages_after_copy,
+)
 
 SectionPage.subpage_types += ['yourwords.YourWordsCompetition']
 
 
-class YourWordsCompetitionIndexPage(Page):
-    parent_page_types = []
+class YourWordsCompetitionIndexPage(Page, PreventDeleteMixin):
+    parent_page_types = ['core.Main']
     subpage_types = ['yourwords.YourWordsCompetition']
 
+    def copy(self, *args, **kwargs):
+        site = kwargs['to'].get_site()
+        main = site.root_page
+        YourWordsCompetitionIndexPage.objects.child_of(main).delete()
+        super(YourWordsCompetitionIndexPage, self).copy(*args, **kwargs)
 
-class YourWordsCompetition(TranslatablePageMixin, Page):
+
+@receiver(index_pages_after_copy, sender=Main)
+def create_yourwords_competition_index_page(sender, instance, **kwargs):
+    if not instance.get_children().filter(
+            title='Your words competitions').exists():
+        yourwords_competition_index = YourWordsCompetitionIndexPage(
+            title='Your words competitions', slug=('yourwords-%s' % (
+                generate_slug(instance.title), )))
+        instance.add_child(instance=yourwords_competition_index)
+        yourwords_competition_index.save_revision().publish()
+
+
+class YourWordsCompetition(TranslatablePageMixinNotRoutable, Page):
+    parent_page_types = [
+        'yourwords.YourWordsCompetitionIndexPage', 'core.SectionPage']
     subpage_types = ['yourwords.TermsAndConditions', 'yourwords.ThankYou']
     description = models.TextField(null=True, blank=True)
     image = models.ForeignKey(
@@ -112,20 +139,49 @@ class YourWordsCompetitionEntry(models.Model):
         help_text=_('Page to which the entry was converted to')
     )
 
+    panels = [
+        MultiFieldPanel(
+            [
+                # TODO: Use ReadOnlyPanel for story_name and story_text
+                # TODO: Add back other fields as read_only
+                FieldPanel('competition'),
+                FieldPanel('story_name'),
+                FieldPanel('story_text'),
+                FieldPanel('is_read'),
+                FieldPanel('is_shortlisted'),
+                FieldPanel('is_winner'),
+            ],
+            heading="Entry Settings",)
+    ]
+
     class Meta:
         verbose_name = 'YourWords Competition Entry'
         verbose_name_plural = 'YourWords Competition Entries'
 
 
 class TermsAndConditions(ArticlePage):
+    parent_page_types = ['yourwords.YourWordsCompetition']
     subpage_types = []
 
     def get_parent_page(self):
         return YourWordsCompetition.objects.all().ancestor_of(self).last()
+
+
+TermsAndConditions.promote_panels = [
+    MultiFieldPanel(
+        Page.promote_panels,
+        "Common page configuration", "collapsible collapsed")]
 
 
 class ThankYou(ArticlePage):
+    parent_page_types = ['yourwords.YourWordsCompetition']
     subpage_types = []
 
     def get_parent_page(self):
         return YourWordsCompetition.objects.all().ancestor_of(self).last()
+
+
+ThankYou.promote_panels = [
+    MultiFieldPanel(
+        Page.promote_panels,
+        "Common page configuration", "collapsible collapsed")]